feat: 자동차가 5자 이하의 이름을 갖도록 구현
자동차(Car)와 기어(CarGear)의 역할 분리를 위해 클래스명 변경.
Car 클래스에서 이름의 길이가 5을 초과하면 Exception을 발생시킴.
