+++ README.md
@@ -31,11 +31,21 @@
 
 ---
 ### 구현 목록
-[ ] 자동차 객체 생성.  
-[ ] 자동차가 이름을 갖도록 개발.  
+- 자동차  
+[X] 자동차 객체 생성.  
+[X] 자동차의 이름은 5자 이하이도록 개발.  
+[ ] 자동차의 이름은 , 로 구분되도록 개발.  
+[X] 자동차가 이름을 갖도록 개발.  
 [ ] 자동차들을 가지는 객체 생성.  
+[ ] 자동차가 위치값 생성.
+  
+- 기어
 [x] random 값 생성 및 검증.     
 [X] random 값이 4이상이면 전진, 3이하면 멈추는 로직 생성.  
-[ ] random 값 결과에 따라 위치를 조정하는 로직 생성.  
-[ ] 결과값 객체 생성.   
+[ ] random 값 결과에 따라 위치를 조정하는 로직 생성.
+  
+- 결과  
+[ ] 결과값 객체 생성.  
+  
+- 사용자 입력  
 [ ] 사용자 입력 값 받기.
