+++ src/main/java/com/nextstep/precourse/racingcar/Car.java
@@ -1,15 +1,13 @@
 package com.nextstep.precourse.racingcar;
 
 public class Car {
+	public static final int CAR_NAME_MAX_LENGTH = 5;
+	private String carName;
 
-	public int getRandomNum() {
-		return (int)(Math.random() * 9 + 1);
-	}
-
-	public CarStatus move(int randomNum) {
-		if (randomNum >= 4) {
-			return CarStatus.FORWARD;
+	public Car(String carName) {
+		if(carName.length() > CAR_NAME_MAX_LENGTH){
+			throw new IllegalArgumentException("자동차 이름은 5자 이하만 가능합니다.");
 		}
-		return CarStatus.STOP;
+		this.carName = carName;
 	}
 }
