+++ src/main/java/com/nextstep/precourse/racingcar/CarGear.java
@@ -1,0 +1,15 @@
+package com.nextstep.precourse.racingcar;
+
+public class CarGear {
+
+	public int getRandomNum() {
+		return (int)(Math.random() * 9 + 1);
+	}
+
+	public CarStatus move(int randomNum) {
+		if (randomNum >= 4) {
+			return CarStatus.FORWARD;
+		}
+		return CarStatus.STOP;
+	}
+}
