+++ src/test/java/com/nextstep/precourse/racingcar/CarTest.java
@@ -1,32 +1,23 @@
 package com.nextstep.precourse.racingcar;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 
 public class CarTest {
-	@Test
-	@DisplayName("랜덤한 숫자 생성되는지 테스트.")
-	void 랜덤숫자_생성_테스트() {
-		Car car = new Car();
-		int number = car.getRandomNum();
-		assertThat(ValidationUtils.validateNumber(number)).isTrue();
+	@ParameterizedTest
+	@ValueSource(strings = {"abcf", "aaaaa"})
+	@DisplayName("자동차의 이름이 6자 이상이면 Exception을 발생시키는지 테스트.")
+	void 자동차이름_5자이하면_생성되는지_테스트(String carName){
+		new Car(carName);
 	}
 
-	@Test
-	@DisplayName("랜덤한 숫자가 4이상이면 전진을 리턴하는지 테스트.")
-	void 랜덤값이_4이상이면_전진(){
-		Car car = new Car();
-		CarStatus carStatus = car.move(4);
-		assertThat(carStatus).isEqualTo(CarStatus.FORWARD);
-	}
-
-	@Test
-	@DisplayName("랜덤한 숫자가 3이하면 멈춤을 리턴하는지 테스트.")
-	void 랜덤값이_3이하면_멈춤(){
-		Car car = new Car();
-		CarStatus carStatus = car.move(3);
-		assertThat(carStatus).isEqualTo(CarStatus.STOP);
+	@ParameterizedTest
+	@ValueSource(strings = {"abcdef", "aaaaaaa"})
+	@DisplayName("자동차의 이름이 6자 이상이면 Exception을 발생시키는지 테스트.")
+	void 자동차이름_6자이상일때_exception_발생하는지_테스트(String carName) {
+		assertThrows(IllegalArgumentException.class, () -> new Car(carName));
 	}
 }
